--- fs/smb/common/smbdirect/Makefile
@@ -1,18 +1,0 @@
-# SPDX-License-Identifier: GPL-2.0-or-later
-#
-# Makefile for smbdirect support
-#
-
-obj-$(CONFIG_SMB_COMMON_SMBDIRECT) += smbdirect.o
-
-smbdirect-y := \
-	smbdirect_socket.o	\
-	smbdirect_connection.o	\
-	smbdirect_mr.o		\
-	smbdirect_rw.o		\
-	smbdirect_debug.o	\
-	smbdirect_connect.o	\
-	smbdirect_listen.o	\
-	smbdirect_accept.o	\
-	smbdirect_devices.o	\
-	smbdirect_main.o
